% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginal_effect.R
\name{print.marginal_cox}
\alias{print.marginal_cox}
\title{Print the marginal treatment effects for hazard ratio (HR)}
\usage{
\method{print}{marginal_cox}(x, ...)
}
\arguments{
\item{x}{an object of class 'marginal_cox'}

\item{...}{Parameters for other methods.}
}
\description{
print method for class 'marginal_cox'.
}
\keyword{internal}
