% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmst_estimate.R
\name{print.rmst_bunsen}
\alias{print.rmst_bunsen}
\title{Print the marginal restricted mean survival time (RMST)}
\usage{
\method{print}{rmst_bunsen}(x, ...)
}
\arguments{
\item{x}{an object of class 'rmst_bunsen'}

\item{...}{Parameters for other methods.}
}
\value{
No return value. This is called for its side effects.
}
\description{
print method for class 'rmst_bunsen'.
}
\keyword{internal}
