% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginal_effect.R
\name{print.summary.marginal_cox}
\alias{print.summary.marginal_cox}
\title{Print the summary of marginal treatment effects for hazard ratio (HR)}
\usage{
\method{print}{summary.marginal_cox}(x, ...)
}
\arguments{
\item{x}{an object of class 'summary_marginal_cox'}

\item{...}{Parameters for other methods.}
}
\value{
No return value. This is called for its side effects.
}
\description{
print method for class 'summary_marginal_cox'.
}
\keyword{internal}
