% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_coxmodel.coxph}
\alias{sanitize_coxmodel.coxph}
\title{Check if the coxph model is correctly specified.}
\usage{
\method{sanitize_coxmodel}{coxph}(model, trt, ...)
}
\arguments{
\item{model}{A coxph model from survival package.}

\item{trt}{Character. Name of the treatment assignment variable.}
}
\description{
Check the covariates of the coxph model.
}
\keyword{internal}
