% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_coxmodel.default}
\alias{sanitize_coxmodel.default}
\title{Check if the model is supported.}
\usage{
\method{sanitize_coxmodel}{default}(model, ...)
}
\arguments{
\item{model}{A coxph model from survival package.}

\item{...}{Parameters for other methods.}
}
\value{
No return value. This is called for its side effects.
}
\description{
At the moment, only coxph is supported for the time-to-event endpoints.
}
\keyword{internal}
