% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_coxmodel}
\alias{sanitize_coxmodel}
\title{Check the coxph model fit and model specifications.}
\usage{
sanitize_coxmodel(model, ...)
}
\arguments{
\item{model}{A coxph model from survival package.}

\item{...}{Parameters for other methods.}
}
\description{
Check the coxph model fit including model class and covariates.
See if the model is supported for estimating the marginal treatment effects.
}
