% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginal_effect.R
\name{summary.marginal_cox}
\alias{summary.marginal_cox}
\title{Summarizing the marginal treatment effects for hazard ratio (HR)}
\usage{
\method{summary}{marginal_cox}(object, ...)
}
\arguments{
\item{object}{an object of class 'marginal_cox'}

\item{...}{Parameters for other methods.}
}
\description{
summary method for class 'marginal_cox'.
}
\keyword{internal}
