% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunsen-package.R
\docType{package}
\name{bunsen-package}
\alias{bunsen}
\alias{bunsen-package}
\title{bunsen: Marginal Survival Estimation with Covariate Adjustment}
\description{
Provides an efficient and robust implementation for estimating marginal Hazard Ratio (HR) and Restricted Mean Survival Time (RMST) with covariate adjustment using Daniel et al. (2021) \doi{10.1002/bimj.201900297} and Karrison et al. (2018) \doi{10.1177/1740774518759281}.
}
\author{
\strong{Maintainer}: Xinlei Deng \email{xinlei.deng@novartis.com} (\href{https://orcid.org/0000-0001-8129-6007}{ORCID})

Authors:
\itemize{
  \item Mark Baillie \email{mark.baillie@novartis.com} (\href{https://orcid.org/0000-0002-5618-0667}{ORCID})
  \item Craig Wang \email{craig.wang@novartis.com} (\href{https://orcid.org/0000-0003-1804-2463}{ORCID})
  \item Dominic Magirr \email{dominic.magirr@novartis.com}
  \item Alex Przybylski \email{alexander.przybylski@novartis.com}
}

}
\keyword{internal}
