% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunsen-package.R
\docType{package}
\name{bunsen-package}
\alias{bunsen}
\alias{bunsen-package}
\title{bunsen: Marginal Estimation with Covariate Adjustment for Survival Endpoint in Clinical Trials}
\description{
A lightweight package aims to provide an easy-to-use interface for estimating marginal or unconditional Hazard Ratio (HR) and Restricted Mean Survival Time (RMST) when adjusting prognostic covariates in clinical trials.
}
\author{
\strong{Maintainer}: Xinlei Deng \email{xinlei.deng@novartis.com} (\href{https://orcid.org/0000-0001-8129-6007}{ORCID})

Authors:
\itemize{
  \item Mark Baillie \email{mark.baillie@novartis.com} (\href{https://orcid.org/0000-0002-5618-0667}{ORCID})
  \item Craig Wang \email{craig.wang@novartis.com} (\href{https://orcid.org/0000-0003-1804-2463}{ORCID})
  \item Dominic Magirr \email{dominic.magirr@novartis.com}
  \item Alex Przybylski \email{alexander.przybylski@novartis.com}
}

}
\keyword{internal}
